+++ meta/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the Orbital Trading project will be documented in this fi
 
 ## [30.50] - 2025-10-24
 - adjusted navigation/market containers
-- 
+- bug fix on 'max' button appearance
+-
 
 ## [30.40] - 2025-10-23
 - fixed license bug